+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+
+## Improved [v1.4.12] 2023-07-17
+- resolve bug in add plot screen
+- ui force update
 ## Enhance [v1.4.11] 2023-07-17
 
 - pin news